am: rewrite IStorage
