+++ src/core/hle/service/am/applet_manager.cpp
@@ -12,6 +12,7 @@
 #include "core/hle/service/am/frontend/applet_controller.h"
 #include "core/hle/service/am/frontend/applet_mii_edit_types.h"
 #include "core/hle/service/am/frontend/applet_software_keyboard_types.h"
+#include "core/hle/service/am/service/storage.h"
 #include "hid_core/hid_types.h"
 
 namespace Service::AM {
