+++ src/core/hle/service/am/frontend/applets.cpp
@@ -26,7 +26,7 @@
 #include "core/hle/service/am/frontend/applet_software_keyboard.h"
 #include "core/hle/service/am/frontend/applet_web_browser.h"
 #include "core/hle/service/am/frontend/applets.h"
-#include "core/hle/service/am/storage.h"
+#include "core/hle/service/am/service/storage.h"
 #include "core/hle/service/sm/sm.h"
 
 namespace Service::AM::Frontend {
